Revert "Refactor: move version.h from source_main to source_base (#7611)"
This reverts commit 8e5d4574622c97b5fca42c9a67a64ad871e45fd2.
